Refactor pre-arrivals admin:
- Extracted components from the existing components
- Moved dedicated states and handlers to the specific components that are not required in parent component
- Refactor table header to write it in constants and use it to map table headers
