+++ frontend/admin/src/components/PreArrivals/ReceiveModal.tsx
@@ -1,6 +1,9 @@
 import React, { useState } from 'react';
-import { Box, Typography, Modal, IconButton, Button, Chip, CircularProgress } from '@mui/material';
-import { Close as CloseIcon } from '@mui/icons-material';
+import { Box, Modal } from '@mui/material';
+
+import ReceiveModalHeader from './ReceiveModalHeader';
+import ReceiveModalContent from './ReceiveModalContent';
+import ReceiveModalActions from './ReceiveModalActions';
 import type { PreArrival } from '../../types/PreArrival';
 
 interface ReceiveModalProps {
@@ -36,7 +39,7 @@ const ReceiveModal: React.FC<ReceiveModalProps> = ({
       open={isOpen}
       onClose={() => {
         if (!isReceiving) {
-            onClose();
+          onClose();
         }
       }} 
       aria-labelledby="customer-details-modal"
@@ -53,96 +56,15 @@ const ReceiveModal: React.FC<ReceiveModalProps> = ({
         boxShadow: 24,
         p: 4,
       }}>
-        {/* Modal Header */}
-        <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 3 }}>
-          <Typography variant="h6" component="h2" sx={{ fontWeight: 600 }}>
-            Receive
-          </Typography>
-          <IconButton onClick={onClose} size="small">
-            <CloseIcon />
-          </IconButton>
-        </Box>
+        <ReceiveModalHeader onClose={onClose} />
 
-        {/* Modal Content */}
-        {selectedItem && (
-          <Box sx={{ mb: 2 }}>
-            <Box sx={{ mb: 1.5 }}>
-              <Typography variant="body2" sx={{ mb: 0.5, color: '#6b7280', fontSize: '0.875rem', fontWeight: 600 }}>
-                Customer
-              </Typography>
-              <Typography variant="body1" sx={{ fontWeight: 700, fontSize: '1rem', color: '#111827' }}>
-                {selectedItem.user} ({selectedItem.suite})
-              </Typography>
-            </Box>
+        <ReceiveModalContent selectedItem={selectedItem} />
 
-            <Box sx={{ mb: 1.5 }}>
-              <Typography variant="body2" sx={{ mb: 0.5, color: '#6b7280', fontSize: '0.875rem', fontWeight: 600 }}>
-                OTP
-              </Typography>
-              <Typography variant="body1" sx={{ fontWeight: 700, fontSize: '1rem', color: '#111827' }}>
-                {selectedItem.otp}
-              </Typography>
-            </Box>
-
-            <Box sx={{ mb: 1.5 }}>
-              <Typography variant="body2" sx={{ mb: 0.5, color: '#6b7280', fontSize: '0.875rem', fontWeight: 600 }}>
-                Tracking / Order No.
-              </Typography>
-              <Typography variant="body1" sx={{ fontWeight: 700, fontSize: '1rem', color: '#111827' }}>
-                {selectedItem.tracking_no || 'N/A'}
-              </Typography>
-            </Box>
-
-            <Box sx={{ mb: 1.5 }}>
-              <Typography variant="body2" sx={{ mb: 0.5, color: '#6b7280', fontSize: '0.875rem', fontWeight: 600 }}>
-                ETA
-              </Typography>
-              <Typography variant="body1" sx={{ fontWeight: 700, fontSize: '1rem', color: '#111827' }}>
-                {selectedItem.estimate_arrival_time}
-              </Typography>
-            </Box>
-
-            <Box sx={{ mb: 1.5 }}>
-              <Typography variant="body2" sx={{ mb: 0.5, color: '#6b7280', fontSize: '0.875rem', fontWeight: 600 }}>
-                Status
-              </Typography>
-              <Chip
-                label={selectedItem.status.charAt(0).toUpperCase() + selectedItem.status.slice(1).toLowerCase()}
-                size="small"
-                sx={{
-                  bgcolor: selectedItem.status.toLowerCase() === 'received' ? '#dcfce7' : '#fef3c7',
-                  color: selectedItem.status.toLowerCase() === 'received' ? '#166534' : '#92400e',
-                  fontWeight: 700,
-                  fontSize: '0.75rem',
-                  height: '24px',
-                  borderRadius: '5px'
-                }}
-              />
-            </Box>
-
-            <Box sx={{ mb: 1.5 }}>
-              <Typography variant="body2" sx={{ mb: 0.5, color: '#6b7280', fontSize: '0.875rem', fontWeight: 600 }}>
-                Package Details
-              </Typography>
-              <Typography variant="body1" sx={{ fontWeight: 700, fontSize: '1rem', color: '#111827' }}>
-                {selectedItem.details || 'N/A'}
-              </Typography>
-            </Box>
-          </Box>
-        )}
-
-        {/* Modal Actions */}
-        <Box sx={{ display: 'flex', justifyContent: 'flex-end' }}>
-          <Button
-            variant="contained"
-            color="primary"
-            onClick={handleReceiveClick}
-            disabled={selectedItem?.status.toLowerCase() === 'received' || isReceiving}
-            sx={{ minWidth: 100, fontWeight: 600, bgcolor: '#8b5cf6','&:hover': { bgcolor: '#7c3aed'}}}
-          >
-            {isReceiving ? <CircularProgress size={24} color="inherit" /> : 'Receive'}
-          </Button>
-        </Box>
+        <ReceiveModalActions
+          selectedItem={selectedItem}
+          isReceiving={isReceiving}
+          onReceiveClick={handleReceiveClick}
+        />
       </Box>
     </Modal>
   );
